[audio] Inspect for min lead time changes
On a per-renderer and per-capturer basis, change 'minimum
lead time' (or fence time for capturers) from a single
static value to a trio of values: initial value, current
value, and time that value changed to current value. If
needed, this could be changed to an ordered map in the
future (if the full history is needed).

This should aid in the diagnosis of render/capture issues
where we observe the client's packet handling getting out-
of-sync with audio_core's expectations (as shown by
renderer underflows).

Test: new unit tests, plus Inspect shows the changes.

Change-Id: I12b9b81125e692d090d4b6c1ccacf220bb9d95f6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/921193
Reviewed-by: Cam Lloyd <camlloyd@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>