[sandbox] Remove flags and target_moniker from Sender/Receiver
Sender and Receiver now transfer only Zircon handles, not Messages.

Flags are not needed because Sender/Receiver are only used to route
protocols, and flags are not used when connecting to a protocol.
Target moniker was used in a previous version of routing code to launch
the target component on connect, but is now unused.

Removes a dep on the moniker lib.

Bug: b/298112397
Change-Id: I6fe067358e9c723daa44ecb93f1dfc96b04a61c8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/925295
Reviewed-by: Gary Bressler <geb@google.com>
Commit-Queue: Yegor Pomortsev <ypomortsev@google.com>