[docs] Remove references to RFC-0083 and add note
This CL changes docs to reference the main "FIDL versioning" page rather
than "RFC-0083: FIDL versioning". It also adds a note to the top of the
RFC linking to the more useful documentation pages.

RFCs are historical documents and not meant to be kept updated, but
people end up on this RFC (hopefully less often with links to it fixed),
so it's useful to point them in the right direction.

Change-Id: Ie06d9cb5bb1852e059f7860123fcc11406cf4601
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/921134
Reviewed-by: Seth Ladd <sethladd@google.com>
Reviewed-by: Ian McKellar <ianloic@google.com>
Commit-Queue: Mitchell Kember <mkember@google.com>