[netstack3] Hold strong DeviceIds in route table
This is possible now that we can await async cleanup of strong DeviceIds
and allows us to simplify routing logic. Before we could await cleanup
of strong DeviceIds, we _could_ have simply removed all routes on a
device before removing the device itself (and synchronously asserting
that all strong references were gone). However, this would have
required us to add some form of locking to the routing table to prevent
new routes from being added before we removed the device.

Now we get that synchronization by having events-in-flight hold weak
DeviceIds to ensure that devices marked for deletion cannot have new
routes installed (as when a device is marked for deletion, weak
references to it can no longer be upgraded).

Routing logic is now simpler, as routes that are in process of being
removed from the table now hold strong DeviceIds that allow us to
retrieve data necessary for notifying FIDL watchers of their
disappearance, even if the routes are being removed because the device
itself is being removed.

Fixed: 132990
Change-Id: Ic1943ddef000a73dedfd5c7df3167e420902b7fd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/922432
Commit-Queue: Evan Wildenhain <wildenhain@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>