[shac] Move clippy logic into shac
The staticlints tool is deprecated in favor of shac, so move over the
clippy logic. Now it will run as part of the `static-checks` builder
instead of Tricium, but will still emit Gerrit comments.

The clippy file-parsing logic is complicated enough that I left it as a
separate Go tool instead of moving it into shac Starlark code. However,
this change removes the restriction that the implementation must be
entirely in Go, so in the future we could delete the clippy-related Go
code entirely and instead have shac call into the same Python logic as
`fx clippy`. That's a bigger and riskier change which is why I didn't do
it here.

Bug: 82386
Change-Id: Ib3fdc427f185c3ef84580290647f89b093a360ac
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/924613
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: Joseph Ryan <josephry@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>