[ffx][compat] Exposing json-help from ffx_tool.
This CL enables using `--machine json` with the `--help` option
to print the JSON serialized help for ffx and all subcommands.

The JSON help is used to create a CLI golden test to detect changes
in the command line interface which may break compatibility.

Bug: b/291958926

Change-Id: I5ae5aeb3ae2c483a8fafadd591a5b02348b35542
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/898903
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Megan Batty <mgnb@google.com>