[pkg] Create PackageResolver-ota capability
This fuchsia.pkg.PackageResolver capability, exposed by the
pkg-resolver, is the same as the original, non -ota,
implementation, except that the resolved packages are not added to the
dynamic index. This capability is only to be used by the
system-updater, which will protect the packages from GC by adding them
to the retained index first.

Later CLs will:
  1. Migrate the system-updater to this new capability.
  2. Disentangle the retained and dynamic indices, by having
     membership in the retained index no longer prevent addition to
     the dynamic index.

At which point the only client of the dynamic index will be non-OTA
package resolves and we can replace it with open package tracking.

Fixed: b/293331949
Change-Id: I3ca1b97d41ca731bc5634c13c511f8992836e9ed
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/921312
API-Review: Chris Suter <csuter@google.com>
Reviewed-by: Laura Peskin <pesk@google.com>
Commit-Queue: Ben Keller <galbanum@google.com>
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>