[kernel][tlb] Fix bug in shared entry invalidation
Fixes a bug in which TLB entries that are part of a shared address
region are not invalidated correctly. The code originally attempted to
use an `invlpg` instruction to invalidate entries of a given address
across all PCIDs. However, this is not actually what the `invlpg`
instruction does; from the intel documentation:

"The instruction invalidates TLB entries associated with the current
PCID and may or may not do so for TLB entries associated with other
PCIDs."

This was leading to certain tests accessing stale TLB entries and
crashing. This CL fixes the issue by upgrading all TLB invalidations of
shared entries to a full TLB shootdown (though it is important to
highlight the fact that this may not always shoot down all global
entries).

Bug: 132980
Change-Id: I018d8dcc147b1d6cf1cd9a56e9b68f960431fc6f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/926019
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Adrian Danis <adanis@google.com>