[scurtiny-x] Drop internal model of component manifests
`scrutiny_x::Component` behaves as a component manifest that stores the
identity of the package the component happened to be loaded from. In
other areas of the API, we have recently decided that the object from
which another object was constructed is information easily retained by
clients that need to maintain such a relationship, and need not be
exposed via a Scrutiny X API. This change aligns management of component
manifests with that philosophy.

Change-Id: I6580503695d7dc4ff146a6da91c594869eb6e1b6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/925093
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Laura Peskin <pesk@google.com>
Fuchsia-Auto-Submit: Mark Dittmer <markdittmer@google.com>