Reland "[kernel][arch] Introduce memory types to paging traits"
This is a reland of commit 247c60cab19a9efedbd2af7c6f11d8679136c232

What changed: PagingSettings was made to be a definition of the traits
themselves, allowing control over the default memory value. The previous
approach of using a std::optional<MemoryType> to allow for a safe
default-constructed value ironically wasn't used safely by the author
and led to unintentional dereferencing of the null state (as well as UX
complications).

Also, the paging trait definitions are moved to the top of each
<lib/arch/$arch/page-table.h>, prioritizing the definition of the main
attraction in that file and reducing the expedient definitions made in
the library scope.

Original change's description:
> [kernel][arch] Introduce memory types to paging traits
>
> This change introduces memory type support to the paging traits. This
> provides the functionality to configure pages as normal memory or MMIO.
>
> Bug: 129344
> Change-Id: I1b552b33bc1e439a5ab2ab86fd53b1458dc8ab99
> Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/878980
> Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
> Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
> Reviewed-by: Roland McGrath <mcgrathr@google.com>

Bug: 129344
Change-Id: I6f835013a3b89022f2192c9ba7171d70a38e1561
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/923652
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>