[driver_manager] Fix flake in restart_on_crash
Fixes a flake in the restart on crash test that occured when the node
closure handler was called later than the FinishRestart function. This
was causing an extra removal to happen. The fix is to reset the
node_controller manually before we try to start back up.

Bug: 134678
Tests: Added coverage for node/binding races in unit test
Change-Id: I617ef3cd2d27b0eef28e74588d60024f55320e5d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/927221
Fuchsia-Auto-Submit: Novin Changizi <novinc@google.com>
Reviewed-by: Sarah Chan <spqchan@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>