[starnix] Don't attempt to sync notifier and notifiee
...of a `PortWaiter`.

In commit 3027cfc2faee116546f84aa918951299ac3f860f (fuchsia CL
https://fuchsia-review.googlesource.com/c/fuchsia/+/924473), it was
discussed whether the `PortWaiter` should specify a happens-before
relationship between a notifier and a notifee. To accomplish this, the
`has_pending_user_packet.swap` operation used Acquire/Release ordering.
However, even with such ordering, the `PortWaiter` cannot guarantee the
happens-before relationship in all cases.

This is because immediately after an atomic load with Acquire ordering
but before the waiter processes any triggered events, a new event can be
triggered and there will be no following atomic load to synchronize any
atomic store (with Release ordering) with. Because of this, the only
way to synchronize a waiter with a notifier is to have the waiter
perform the necessary synchronization itself.

Change-Id: If58fd2c757ad32e97f7cdddb6b635a13c580b36c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/927213
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Ghanan Gowripalan <ghanan@google.com>