[amlogic-display][vout] Ignore `mode` if EDID is not present.
Per definition of `edid_present` in `AddedDisplayArgs`, for a given
display, if its EDID information is not present, the meaning of
the `mode` field of DisplayConfig struct will be undefined and the
driver should ignore it in CheckConfiguration() and
ApplyConfiguration().

This change adds a field `ignore_display_mode_` to AmlogicDisplay
so that it can ignore the `mode` field when checking or applying
configurations.

Test: no-op expected.
Test: display-tool squares 1 on Nelson (MIPI-DSI)
Test: display-tool squares 1 on VIM3 (w/ MIMO HDMI display)
Change-Id: Ie4759a76caeb458f678c324ef8120b9b43ddb373
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/924397
Reviewed-by: Victor Costan <costan@google.com>
Fuchsia-Auto-Submit: Yilong Li <liyl@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>