[bazel-sdk]  add ability to set the fuchsia api level
This gives us the ability to set the fuchsia api level for a package.
The package will then transition the command line flags used to stamp
packages, compile fidl libraries and compile c++ files. We provide a way
to set this at the package level and on the command line.

BUG: b/299337628
Change-Id: I0a9b5cc6b3a3745c621ce75ade5c72466d04c258
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/926337
Reviewed-by: Darren Chan <chandarren@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>