[ffx][compat] Fixing golden file generation
This is fixing the golden file generation when a new
sub command is added.

The problem is that the loop creating the golden comparison targets was missing dependencies on the sdk - where ffx comes from so it would use the contents of the SDK before it was finished building the sdk, picking up the old command info.
so now the sdk has to be built, then the top level command list, then iterate over that list, vs. using a stale list.

Bug: b/303542110

Change-Id: I4c117d260648ab899edd876138219ac97648ffca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/927835
Reviewed-by: Megan Batty <mgnb@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>