[bazel-sdk] Add a repository default api level flag
Adds the ability to set a flag in a .bazelrc file which will serve as
a fallback api level if one is not set on a package or on the command
line.

This CL adds 1 new flag and updates a previous flag. Each of these flags
will be put behind a flag alias to make them easier to use.

--@fuchsia_sdk//fuchsia:override_fuchsia_api_level allows a user to
override the api level for a single command.
--@fuchsia_sdk//fuchsia:repository_default_fuchsia_api_level allows for
a fallback api level to set on a global level.

BUG: b/303684806
Change-Id: If3fa9e0b9ffaf4748ec7623ac0926c6cf884cad7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/927841
Reviewed-by: Jiaming Li <lijiaming@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>