[bringup][assembly] Remove special-casing to add base drivers to bootfs
This removes the special logic that copies the contents of all base
drivers into bootfs on a bringup build.

The device-enumeration test is changed to include the packaged drivers
when compiling the test package, but not when compiling the bootfs
test.

Change-Id: I76c8b92170505cf24d144c20ad5eb8485b419ddb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/927844
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>
Reviewed-by: Novin Changizi <novinc@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>