[ffx] add net diagnostics request to e2e_emu tests
A flake that seems to be related to netstack can't be diagnosed
with normal, networking-based tools. Instead we'll add a temporary
hack for a set of tests that seem to exercise the failure,
which invoke some target-side diagnostics that can be inspected
by the netstack team.

BYPASS_INCLUSIVE_LANGUAGE_REASON=TERM:dumb is required
Bug: fxb/133234
Change-Id: Id756955107760a6e7f083fbcb22e20e1fafa3e71

Change-Id: I98b53a4ad2fcc47d5d4514e9f477d19298c1d7cd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/928334
Commit-Queue: Steven Grady <slgrady@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>