[ffx][testing]: better checks on test emulators
Adds macro assert_virtualization_available to check if virtualization is
avaialable on the running platform. Panics if not

Adds field emulator_allowed to the TestContext struct.

Adds new test fixture emulator_fixture, which asserts that
virtualization is available and sets the TestContext to pass to the test
to allow virtualization.

Emu's start function now checks if the passed TestContext allows
virtualization and panics if not.

Change-Id: Ia7a60e62863584bd4b072936b8deaee508fce9b4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/925537
Reviewed-by: Megan Batty <mgnb@google.com>
Commit-Queue: Colin Nelson <colnnelson@google.com>