[ffx][resolvers]: Make Resolver not require flash manifest
Resolver now takes a root_path and does not implement the
ManifestResolver trait.

Added new struct FlashManifestResolver in the manifest mod which embeds
the Resolver and implements the FileResolver and ManifestResolver traits.

It's `new` function asserts the path passed to it is a file.

Change-Id: I9d3e6025820d4c79b8a7ac85f6993b34a0d3146e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/927833
Reviewed-by: Megan Batty <mgnb@google.com>
Commit-Queue: Colin Nelson <colnnelson@google.com>