[rcs] Support connecting to incoming namespace of core/toolbox
This change fixes a bug in remote-control which prevented the ability
of connecting to protocols in the incoming namespace of a component due
to protocols being nested under a svc directory inside the incoming
namespace.

When using OpenDirType::Namespace, the namespace directory looks as:

```
ns/
  svc/
     fuchsia.my.Protocol
  pkg/
```

So requests come for `svc/fuchsia.my.Protocol`.

Additionally this change introduces an API in the rcs library to connect
to protocols in the incoming namespace of core/toolbox.

Fixed: b/304302408
Change-Id: Ia9b429f79173e95127f48bfd3a04d05b223585e7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/929192
Reviewed-by: Megan Batty <mgnb@google.com>
Commit-Queue: Miguel Flores <miguelfrde@google.com>
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>