[vim3] incorporate common/ar64-common and flatten remaining items
This flattens the contents of //boards/common/arm64-common.gni into
vim3.gni, and removes any +/- logic that was a result of the basing
of vim3 on arm64-common.

The remaining items are re-arranged to make clear which are left to
move into assembly, and which are redundant with inputs that are in
assembly, but which are still used by targets in the build.

Bug: b/302567529
Change-Id: I8027aea162712d783a88bd77364db0c9ab2a8cf2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/928464
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>