[starnix][fuse] Workaround b/297439724
Instead of trying to block the system task when closing or flushing a
fuse file, we create a new kernel task and block that task instead. This
works around the `assert` that is triggered when multiple threads try to
write to the run state for the system task in an inconsistent way.

Bug: b/297439724
Change-Id: Ib20a2dcfe4fcceb8747af0badbfdf802ba1606cf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/927174
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Reviewed-by: Benjamin Lerman <qsr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>