[devicetree] Translate address based on ranges
ranges property in the parent node contains address mapping. This needs
to be taken into account when calculating the mmio address.

This CL also fixes reg property parsing for nodes in which reg is not an
mmio. This is currently decided based on the size-cells.

Fixed: 133156
Test: mmio-visitor-test
Change-Id: I0af0bc2a16df1d01e099c49243b428a79e7a7c2b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/924020
Fuchsia-Auto-Submit: Puneetha Ramachandra <puneetha@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Braden Kell <bradenkell@google.com>