[docs] Fixing reference links
This fixes markdown reference links (links with [] that are defined
later in the file).

This is in preparation to turn on doc-checker checking of reference
links, so it will be harder to write updates that have broken links.

Bug: b/289246762
Change-Id: Idbf79db2bde2444474992b69c1d6ab0c2a4b4798
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/930292
Fuchsia-Auto-Submit: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>