[bt][avrcp] CT uses GetItemAttributes when browsed player is set.
According to the spec, controllers should use `GetItemAttributes` browse
command instead of `GetElementAttributes` control command if browsing is
supported on target side. Change the controller service's
`GetMediaAttributes` method, so that the above is implemented. As said
in the spec, the CT registers to uses the UID of the currently playing
track and use in the scope of the Now Playing folder.

Change-Id: Ibcaa0a66cc7b63ea58de1d3d73e0bc5ee6965b62
Bug: 105465
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/845116
Reviewed-by: Ani Ramakrishnan <aniramakri@google.com>
Reviewed-by: Marie Janssen <jamuraa@google.com>
Commit-Queue: Dayeong Lee <dayeonglee@google.com>