[device-enumeration-test] Ignore xhci root hub on vim3
We've seen many (12+) devices that often fail the device enumeration
tests because the XHCI hub doesn't enumerate.  But we don't actually
make use of the XHCI bus in any way, and we'd rather not have to either
tolerate this tests constantly flaking nor do we want to have to pull
the vim3s from the hardware fleet.  So just ignore when this device
fails to enumerate instead for now, and we can revisit if we start doing
anything with the XHCI bus.

Bug: b/296738636
Change-Id: I1089279762c722ce9978f9b5c8111ae9195b401e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/930552
Reviewed-by: Sam Hansen <hansens@google.com>
Fuchsia-Auto-Submit: Drew Fisher <zarvox@google.com>
Commit-Queue: Drew Fisher <zarvox@google.com>